- use "reset" events to modify the robot initial_states, and convert the RobotCfg back to default - move surface_gripper_actions out of binary_joint_actions to a standalone file
