+++ source/isaaclab_assets/isaaclab_assets/robots/universal_robots.py
@@ -3,10 +3,6 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
-# Copyright (c) 2022-2025, The Isaac Lab Project Developers.
-# All rights reserved.
-#
-# SPDX-License-Identifier: BSD-3-Clause
 
 """Configuration for the Universal Robots.
 
@@ -38,10 +34,10 @@ UR10_CFG = ArticulationCfg(
     init_state=ArticulationCfg.InitialStateCfg(
         joint_pos={
             "shoulder_pan_joint": 0.0,
-            "shoulder_lift_joint": -1.5707,
-            "elbow_joint": 1.5707,
-            "wrist_1_joint": -1.5707,
-            "wrist_2_joint": -1.5707,
+            "shoulder_lift_joint": -1.712,
+            "elbow_joint": 1.712,
+            "wrist_1_joint": 0.0,
+            "wrist_2_joint": 0.0,
             "wrist_3_joint": 0.0,
         },
     ),
@@ -72,10 +68,10 @@ UR10_LONG_SUCTION_CFG = ArticulationCfg(
     init_state=ArticulationCfg.InitialStateCfg(
         joint_pos={
             "shoulder_pan_joint": 0.0,
-            "shoulder_lift_joint": -1.5707,
-            "elbow_joint": 1.5707,
-            "wrist_1_joint": -1.5707,
-            "wrist_2_joint": -1.5707,
+            "shoulder_lift_joint": -1.712,
+            "elbow_joint": 1.712,
+            "wrist_1_joint": 0.0,
+            "wrist_2_joint": 0.0,
             "wrist_3_joint": 0.0,
         },
     ),
@@ -104,10 +100,10 @@ UR10_SHORT_SUCTION_CFG = ArticulationCfg(
     init_state=ArticulationCfg.InitialStateCfg(
         joint_pos={
             "shoulder_pan_joint": 0.0,
-            "shoulder_lift_joint": -1.5707,
-            "elbow_joint": 1.5707,
-            "wrist_1_joint": -1.5707,
-            "wrist_2_joint": -1.5707,
+            "shoulder_lift_joint": -1.712,
+            "elbow_joint": 1.712,
+            "wrist_1_joint": 0.0,
+            "wrist_2_joint": 0.0,
             "wrist_3_joint": 0.0,
         },
     ),
